refactor(model): rename client_responses module to client_types
The module also holds types used as builder inputs (SuggestedParams
is consumed by every Builder::build(&params)) and the TransactionParams
trait — neither of which is a 'response'. Future D3 slices will keep
adding more types here, so widen the name now before the misnomer
locks in via callers.
