+++ algonaut_model/src/lib.rs
@@ -1,3 +1,3 @@
-pub mod client_responses;
+pub mod client_types;
 pub mod kmd;
 pub mod transaction;
