+++ src/atomic_transaction_composer/mod.rs
@@ -144,7 +144,7 @@ pub struct AddMethodCallParams {
     /// The address of the sender of this application call
     pub sender: Address,
     /// Transactions params to use for this application call
-    pub suggested_params: algonaut_model::client_responses::SuggestedParams,
+    pub suggested_params: algonaut_model::client_types::SuggestedParams,
     /// The OnComplete action to take for this application call
     pub on_complete: ApplicationCallOnComplete,
     /// The approval program for this application call. Only set this if this is an application
