+++ tests/step_defs/unit/world.rs
@@ -58,11 +58,11 @@ pub enum UnitResponse {
         algonaut::openapi_algod::models::GetPendingTransactionsByAddress200Response,
     ),
     RawTransaction(algonaut::openapi_algod::models::RawTransaction200Response),
-    Status(algonaut_model::client_responses::NodeStatus),
-    Supply(algonaut_model::client_responses::Supply),
+    Status(algonaut_model::client_types::NodeStatus),
+    Supply(algonaut_model::client_types::Supply),
     Account(algonaut::openapi_algod::models::Account),
     Block(algonaut::openapi_algod::ext::block::BlockResponse),
-    TransactionParams(algonaut_model::client_responses::SuggestedParams),
+    TransactionParams(algonaut_model::client_types::SuggestedParams),
     Dryrun(algonaut::openapi_algod::models::TealDryrun200Response),
     AssetBalances(algonaut::openapi_indexer::models::LookupAssetBalances200Response),
     Transactions(algonaut::openapi_indexer::models::LookupAccountTransactions200Response),
